Split Kokkos_SYCL_Parallel_Range.hpp
